refactor(dcc): carve speculative regalloc codegen into dcc_regalloc.c (round-2 item 6)
Move the ~1330-line speculative no-IX-frame and BC-register-allocation
codegen cluster out of dcc_func.c into a new dcc_regalloc.c translation
unit. Promote the 6 entry points plus find_bc_regalloc_candidate and
plain_static_body_can_be_buffered to non-static with declarations in
dcc.h. dcc_func.c: 5115 -> 3749 lines.

Behavior-preserving: 0 asm diffs across 316 tests, 0 diag diffs across
106 diagnostics; fast suite 307 passed / 0 failed / 9 skipped.
